refactor(ledger): use the upstream v8->v9 state translation crate
The v8->v9 state translation table lived in-repo as a hand-port of
midnight-ledger's `v8-to-v9-state-translation` crate, so ledger-side fixes
had to be re-ported by hand and we could drift from downstream consumers
(indexer) that use the real crate.

Replace the 734-line copy with a git dependency on the crate itself. It is
unpublished and not in any release tag - it lives only on the ledger repo's
`state-translation/v8-to-v9` branch - so it is pinned by rev. Both consumers,
the on-chain migration (`host_api::migration_8_to_9`) and the toolkit fork
boundary (`fork::fork_8_to_9`), import it directly rather than through a
re-export shim.

No behaviour change. Upstream is identical to the vendored copy including the
dust wipe, which landed upstream in midnight-ledger PR #707
(`dust: recast(&source.dust)?` -> `dust: Sp::new(Default::default())`) and
closes the TODO the vendored file carried.

Three guard tests move into `migration_8_to_9`'s test module rather than
dying with the file, because they guard *this* workspace's pins rather than
upstream's:

- `table_tags_match_types` - the imported TABLE hardcodes tag literals that
  upstream validates against its own dependency pins, not our
  `[patch.crates-io]` rc pins. An rc bump moving a `#[tag]` would otherwise
  mis-dispatch the migration silently.
- `table_is_closed` - table closure under our crate instances.
- `dust_state_is_wiped` - upstream has no dust test, and
  `pallet_cnight_observation::migrations::v2` is built entirely on the wipe.

`onchain-state-ledger-8`/`-9` drop out of `ledger/helpers`, which only needed
them to name the vendored file's v8/v9 aliases.
