+++ ledger/helpers/Cargo.toml
@@ -23,16 +23,17 @@ reqwest = { workspace = true }
 
 mn-ledger-8 = { workspace = true, features = ["proving"] }
 onchain-runtime-ledger-8 = { workspace = true }
-onchain-state-ledger-8 = { workspace = true }
 ledger-storage-ledger-8 = { workspace = true }
 zswap-ledger-8 = { workspace = true }
 
 mn-ledger-9 = { workspace = true, features = ["proving", "test-utilities"] }
 onchain-runtime-ledger-9 = { workspace = true }
-onchain-state-ledger-9 = { workspace = true }
 zswap-ledger-9 = { workspace = true }
 coin-structure-ledger-9 = { workspace = true }
 transient-crypto-ledger-9 = { workspace = true }
+
+# v8->v9 state translation table, used by the toolkit fork boundary (`fork::fork_8_to_9`).
+v8-to-v9-state-translation = { workspace = true }
 
 rand = { version = "0.8.4", features = ["small_rng", "getrandom"] }
 rayon.workspace = true
