+++ ledger/src/host_api/migration_8_to_9.rs
@@ -13,8 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Host-side v8 -> v9 ledger state translation, driven by
-//! [`crate::state_translation_v8_to_v9::StateTranslationTable`].
+//! Host-side v8 -> v9 ledger state translation, driven by the ledger team's
+//! [`v8_to_v9_state_translation::StateTranslationTable`].
 //!
 //! The on-chain pallet stores only the arena root of the ledger state
 //! (`pallet_midnight::StateKey`, a `tagged_serialize`d `TypedArenaKey<Ledger>`);
@@ -33,7 +33,7 @@
 use crate::ledger_8::api::Ledger as Ledger8;
 use crate::ledger_9::api::Ledger as Ledger9;
 use crate::ledger_9::types::{DeserializationError, LedgerApiError, SerializationError};
-use midnight_node_ledger_helpers::state_translation_v8_to_v9::StateTranslationTable;
+use v8_to_v9_state_translation::StateTranslationTable;
 
 use base_crypto::cost_model::CostDuration;
 use ledger_storage_ledger_8 as storage;
@@ -184,6 +184,103 @@ pub fn migrate_state_v8_to_v9<D: DB>(
 mod tests {
 	use super::*;
 	use ledger_storage_ledger_8::db::InMemoryDB;
+	use midnight_serialize::Tagged;
+	use std::borrow::Cow;
+	use std::ops::Deref;
+	use storage::state_translation::TranslationTable;
+
+	/// Drive the imported translation table to completion over an in-memory arena.
+	fn translate_to_completion(v8: LedgerState8<InMemoryDB>) -> LedgerState9<InMemoryDB> {
+		let tl = TypedTranslationState::<
+			LedgerState8<InMemoryDB>,
+			LedgerState9<InMemoryDB>,
+			StateTranslationTable,
+			InMemoryDB,
+		>::start(Sp::new(v8))
+		.expect("failed to start translation");
+
+		let finished = tl
+			.run(CostDuration::from_picoseconds(1_000_000_000_000))
+			.expect("translation failed");
+
+		finished
+			.result()
+			.expect("failed to get result")
+			.expect("translation did not complete")
+			.deref()
+			.clone()
+	}
+
+	/// Every `TranslationId` a table entry requires must itself be in the table,
+	/// or translation errors at runtime the first time the entry is needed.
+	#[test]
+	fn table_is_closed() {
+		<StateTranslationTable as TranslationTable<InMemoryDB>>::assert_closure();
+	}
+
+	/// The imported `TABLE` hardcodes tag string literals, and upstream validates them
+	/// against *its* dependency pins — not this workspace's `[patch.crates-io]` rc pins.
+	/// If a tag on either the v8 or v9 side drifts (e.g. an rc bump changes a `#[tag]`),
+	/// the literal no longer matches what `T::tag()` produces here and the migration
+	/// silently mis-dispatches. Rebuild every expected ID from the node's actual crate
+	/// types and compare.
+	#[test]
+	fn table_tags_match_types() {
+		use storage::merkle_patricia_trie::{MerklePatriciaTrie, Node};
+		use storage::storable::SizeAnn;
+
+		type V8Ann = mn_ledger_8::annotation::NightAnn;
+		type V9Ann = mn_ledger_9::annotation::NightAnn;
+		type V8Contract = onchain_state_ledger_8::state::ContractState<InMemoryDB>;
+		type V9Contract = onchain_state_ledger_9::state::ContractState<InMemoryDB>;
+
+		let expected: Vec<(Cow<'static, str>, Cow<'static, str>)> = vec![
+			(LedgerState8::<InMemoryDB>::tag(), LedgerState9::<InMemoryDB>::tag()),
+			(
+				mn_ledger_8::structure::LedgerParameters::tag(),
+				mn_ledger_9::structure::LedgerParameters::tag(),
+			),
+			(V8Contract::tag(), V9Contract::tag()),
+			(
+				MerklePatriciaTrie::<V8Contract, InMemoryDB, V8Ann>::tag(),
+				MerklePatriciaTrie::<V9Contract, InMemoryDB, V9Ann>::tag(),
+			),
+			(
+				Node::<V8Contract, InMemoryDB, V8Ann>::tag(),
+				Node::<V9Contract, InMemoryDB, V9Ann>::tag(),
+			),
+			(u128::tag(), u128::tag()),
+			(
+				MerklePatriciaTrie::<u128, InMemoryDB, SizeAnn>::tag(),
+				MerklePatriciaTrie::<u128, InMemoryDB, V9Ann>::tag(),
+			),
+			(Node::<u128, InMemoryDB, SizeAnn>::tag(), Node::<u128, InMemoryDB, V9Ann>::tag()),
+		];
+
+		let actual: Vec<_> = <StateTranslationTable as TranslationTable<InMemoryDB>>::TABLE
+			.iter()
+			.map(|(id, _)| (id.0.clone(), id.1.clone()))
+			.collect();
+
+		assert_eq!(actual, expected);
+	}
+
+	/// The translation wipes dust: whatever generation/utxo state v8 held, the v9 side
+	/// comes out as the empty state genesis starts from. `pallet_cnight_observation`'s
+	/// dust re-apply migration is built entirely on this being true, and upstream has no
+	/// test for it.
+	#[test]
+	fn dust_state_is_wiped() {
+		let mut v8 = LedgerState8::<InMemoryDB>::new("test-network");
+		let mut dust = (*v8.dust).clone();
+		dust.generation.generating_tree_first_free = 7;
+		dust.utxo.commitments_first_free = 3;
+		v8.dust = Sp::new(dust);
+
+		let v9 = translate_to_completion(v8);
+
+		assert_eq!(*v9.dust, mn_ledger_9::dust::DustState::default());
+	}
 
 	/// Dev-only: verify that seeding a v8 genesis blob with *this* crate's
 	/// ledger_8 `Ledger` wrapper reproduces the exact arena root a ledger-8 node
